services/evaluations/requirements.txt
bigframes
functions-framework
google-cloud-aiplatform[evaluation]
google-cloud-bigquery
pandas
pandas-io
pandas-gbq